Moved Movie objects to Movie class
