+++ src/it/scala/com/phasmidsoftware/table/MovieFuncSpec.scala
@@ -29,37 +29,6 @@ class MovieFuncSpec extends AnyFlatSpec with Matchers {
     mt take 10 foreach println
   }
 
-  // TODO move this into Movie
-  // CONSIDER removing the csvAttributes parameter and making it an object.
-  class CsvRendererMovie(implicit val csvAttributes: CsvAttributes) extends CsvRenderers with CsvRenderer[Movie] {
-
-    import com.phasmidsoftware.render.CsvGenerators._
-
-    private val csvGenerators = new CsvGenerators {}
-    implicit val generatorStringList: CsvGenerator[StringList] = csvGenerators.sequenceGenerator[String]
-    implicit val generatorOptionDouble: CsvGenerator[Option[Double]] = csvGenerators.optionGenerator
-    implicit val generatorOptionInt: CsvGenerator[Option[Int]] = csvGenerators.optionGenerator
-    implicit val generatorOptionString: CsvGenerator[Option[String]] = csvGenerators.optionGenerator
-
-    import com.phasmidsoftware.render.CsvRenderers._
-
-    implicit val rendererStringList: CsvRenderer[StringList] = sequenceRenderer[String]
-    implicit val rendererOptionDouble: CsvRenderer[Option[Double]] = optionRenderer
-    implicit val rendererOptionInt: CsvRenderer[Option[Int]] = optionRenderer
-    implicit val rendererOptionString: CsvRenderer[Option[String]] = optionRenderer
-    implicit val rendererFormat: CsvProduct[Format] = rendererGenerator4(Format)
-    implicit val rendererProduction: CsvRenderer[Production] = renderer4(Production)
-    implicit val rendererRating: CsvRenderer[Rating] = renderer2(Rating.apply)
-    implicit val rendererReviews: CsvRenderer[Reviews] = renderer7(Reviews)
-    implicit val rendererName: CsvRenderer[Name] = renderer4(Name.apply)
-    implicit val rendererPrincipal: CsvRenderer[Principal] = renderer2(Principal)
-    implicit val rendererOptionPrincipal: CsvRenderer[Option[Principal]] = optionRenderer
-    val fAttributeSet: StringList => AttributeSet = AttributeSet.apply
-    implicit val rendererAttributeSet: CsvRenderer[AttributeSet] = renderer1(fAttributeSet)
-
-    def render(t: Movie, attrs: Map[String, String]): String = renderer11(Movie).render(t, attrs)
-  }
-
   /**
    * NOTE: it is perfectly proper for there to be a number of parsing problems.
    * These are application-specific and are not indicative of any bugs in the
@@ -71,7 +40,7 @@ class MovieFuncSpec extends AnyFlatSpec with Matchers {
     val mti: IO[Table[Movie]] = IOUsing(Source.fromURL(classOf[Movie].getResource("movie_metadata.csv")))(x => Table.parseSource(x))
     // TODO create a combined renderer/generator for Movie
     implicit val csvRenderer: CsvRenderer[Movie] = new CsvRendererMovie
-    implicit val csvGenerator: CsvGenerator[Movie] = createMovieCvsGenerator
+    implicit val csvGenerator: CsvGenerator[Movie] = Movie.createMovieCvsGenerator
 
     val wi: IO[String] = for (mt <- mti) yield mt.toCSV
     wi.unsafeRunSync().startsWith(
@@ -85,25 +54,5 @@ class MovieFuncSpec extends AnyFlatSpec with Matchers {
     implicit val hasKey: HasKey[Movie] = (t: Movie) => t.production.country
     val wi: IO[Int] = for (mt <- mti) yield mt.filterByKey(_ == "New Zealand").size
     wi.unsafeRunSync() shouldBe 4
-  }
-
-  // TODO move this into Movie
-  private def createMovieCvsGenerator: CsvGenerator[Movie] = {
-    import com.phasmidsoftware.render.CsvGenerators._
-    val csvGenerators = new CsvGenerators {}
-    implicit val generatorStringList: CsvGenerator[StringList] = csvGenerators.sequenceGenerator[String]
-    implicit val generatorOptionDouble: CsvGenerator[Option[Double]] = csvGenerators.optionGenerator
-    implicit val generatorOptionInt: CsvGenerator[Option[Int]] = csvGenerators.optionGenerator
-    implicit val generatorOptionString: CsvGenerator[Option[String]] = csvGenerators.optionGenerator
-    implicit val generatorFormat: CsvGenerator[Format] = csvGenerators.generator4(Format)
-    implicit val generatorProduction: CsvGenerator[Production] = csvGenerators.generator4(Production)
-    implicit val generatorRating: CsvGenerator[Rating] = csvGenerators.generator2(Rating.apply)
-    implicit val generatorReviews: CsvGenerator[Reviews] = csvGenerators.generator7(Reviews)
-    implicit val generatorName: CsvGenerator[Name] = csvGenerators.generator4(Name.apply)
-    implicit val generatorPrincipal: CsvGenerator[Principal] = csvGenerators.generator2(Principal)
-    implicit val generatorOptionPrincipal: CsvGenerator[Option[Principal]] = csvGenerators.optionGenerator
-    val fAttributeSet: StringList => AttributeSet = AttributeSet.apply
-    implicit val generatorAttributeSet: CsvGenerator[AttributeSet] = csvGenerators.generator1(fAttributeSet)
-    csvGenerators.generator11(Movie)
   }
 }
